move files into own directory
