+++ src/layout/Layout.tsx
@@ -10,6 +10,7 @@ import { Footer } from 'layout/Footer/Footer'
 import ScrollToTop from 'components/ScrollToTop/ScrollToTop'
 import { useLocation } from 'react-router-dom'
 import { storeUserProject } from 'utils/helpers/stateHelpers'
+import { ProjectPortal } from 'screens/Landing'
 
 type Props = {
   children: React.ReactNode
@@ -20,6 +21,9 @@ export const Layout: React.FC<Props> = ({ children }: Props) => {
   const dispatch = useAppDispatch()
   const status = useAppSelector(uiStatus)
   const userId = useAppSelector(state => state.ui.auth.user._id)
+  const projectPortal = useAppSelector(
+    state => state.project.projectPortal.renderProjectPortal
+  )
 
   useEffect(() => {
     const verifyUser = async () => {
@@ -42,15 +46,19 @@ export const Layout: React.FC<Props> = ({ children }: Props) => {
       <Nav />
       {userId && <Sidebar />}
       <div className={userId ? 'layout-container active' : ''}>
-        <div
-          className={
-            location.pathname !== '/'
-              ? 'main-content-container'
-              : 'landing-main-content-container'
-          }
-        >
-          {children}
-        </div>
+        {projectPortal ? (
+          <ProjectPortal />
+        ) : (
+          <div
+            className={
+              location.pathname !== '/'
+                ? 'main-content-container'
+                : 'landing-main-content-container'
+            }
+          >
+            {children}
+          </div>
+        )}
       </div>
       <Footer />
     </>
