+++ src/screens/Landing/Landing.tsx
@@ -9,14 +9,10 @@ import { selectAuthUser } from 'utils/redux/slices/userSlice'
 import { getRandomInt } from 'screens/AccountSettings/helper/data'
 import { storeUserProject } from 'utils/helpers/stateHelpers'
 import './Landing.scss'
-import { ProjectPortal } from './ProjectPortal'
 
 export const Landing: React.FC = () => {
   const [loginStatus, setLoginStatus] = useState<boolean | null>(null)
   const authUser = useAppSelector(selectAuthUser)
-  const projectPortal = useAppSelector(
-    state => state.project.projectPortal.renderProjectPortal
-  )
 
   const dispatch = useDispatch()
 
@@ -56,29 +52,25 @@ export const Landing: React.FC = () => {
 
   return (
     <div className='landing-container'>
-      {projectPortal ? (
-        <ProjectPortal />
-      ) : (
-        <div className='header-container'>
-          <div className='header-grid'>
-            <h1>Surpass Your Competition In The Tech Job Market</h1>
-            <p>
-              A platform to collaborate on real-world projects! Don't wait to
-              build your development experience.
-            </p>
-            <Link to='/sign-up'>Start Today!</Link>
-          </div>
-          <div className='developer-grid'>
-            <h1> For UX & Developers Only! </h1>
-            <h2>
-              Login as a random user using the button below <br /> Button is for
-              devs who want to skip auth user flow
-            </h2>
-            <button onClick={randomUserLogin}>Login as random user</button>
-            <LoginStatusSymbol />
-          </div>
+      <div className='header-container'>
+        <div className='header-grid'>
+          <h1>Surpass Your Competition In The Tech Job Market</h1>
+          <p>
+            A platform to collaborate on real-world projects! Don't wait to
+            build your development experience.
+          </p>
+          <Link to='/sign-up'>Start Today!</Link>
         </div>
-      )}
+        <div className='developer-grid'>
+          <h1> For UX & Developers Only! </h1>
+          <h2>
+            Login as a random user using the button below <br /> Button is for
+            devs who want to skip auth user flow
+          </h2>
+          <button onClick={randomUserLogin}>Login as random user</button>
+          <LoginStatusSymbol />
+        </div>
+      </div>
     </div>
   )
 }
