+++ src/screens/Landing/ProjectPortal/ProjectPortal.scss
@@ -1,0 +1,5 @@
+.portal-container {
+  min-height: calc(100vh - 250px);
+  padding: 130px 20px;
+  transition: all 0.1s;
+}
